refactored frontend files
