+++ frontend/src/__tests__/UserProfile.test.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';  // Import BrowserRouter
-import UserProfile from '../UserProfile';
+import UserProfile from '../user/UserProfile';
 
 test('renders user profile with dummy data', () => {
   const dummyUser = {
