+++ frontend/src/components/nav_bar/NavigationBar.css
