+++ frontend/src/events/wellfare_events/WellfarePage.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import './WellfarePage.css';
 import { Link } from "react-router-dom";
-import NavigationBar from './NavigationBar';
-import EventCard from './EventCard'; // Import the EventCard component
+import NavigationBar from '../../components/nav_bar/NavigationBar';
+import EventCard from '../EventCard'; // Import the EventCard component
 
 const WellfarePage = ({ handleLogout }) => {
   const [events, setEvents] = useState([]);
