+++ frontend/src/friends/FriendCard.css
