+++ frontend/src/friends/FriendCard.jsx
